federated_ckd/hospital_B/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
psycopg2-binary   # postgres
